Changed type of 'id' in WordInfoEntity to long and refactored logic that uses it

+++ app/src/main/java/com/peeranm/simpledictionary/feature_word_meaning/data/repository/WordInfoRepository.kt
@@ -5,7 +5,7 @@ import com.peeranm.simpledictionary.core.Resource
 import kotlinx.coroutines.flow.Flow
 
 interface WordInfoRepository {
-    suspend fun getWordInfoById(id: Int): WordInfo
+    suspend fun getWordInfoById(id: Long): WordInfo
     fun getWordInfo(word: String, languageCode: String): Flow<Resource<List<WordInfo>>>
     fun getAllWordInfo(): Flow<List<WordInfo>>
