+++ app/src/main/java/com/peeranm/simpledictionary/feature_word_meaning/use_cases/GetWordInfoFromCacheUseCase.kt
@@ -2,10 +2,10 @@ package com.peeranm.simpledictionary.feature_word_meaning.use_cases
 
 import com.peeranm.simpledictionary.feature_word_meaning.data.repository.WordInfoRepository
 import com.peeranm.simpledictionary.feature_word_meaning.model.WordInfo
-import kotlinx.coroutines.flow.Flow
+
 
 class GetWordInfoFromCacheUseCase(private val repository: WordInfoRepository) {
-    suspend operator fun invoke(id: Int): WordInfo {
+    suspend operator fun invoke(id: Long): WordInfo {
        return repository.getWordInfoById(id)
     }
 }
